Pages test di

+++ Models/LoaiDaiLy.cs
@@ -11,7 +11,7 @@ namespace QuanLyDaiLy.Models
         [Required(ErrorMessage = "Tên loại đại lý không được để trống")]
         [StringLength(100, ErrorMessage = "Tên loại đại lý không được vượt quá 100 ký tự")]
         public string TenLoaiDaiLy { get; set; } = "";
-        public int NoToiDa { get; set; } = 0;
+        public long NoToiDa { get; set; } = 0;
 
         // Navigation property
         public ICollection<DaiLy> DsDaiLy { get; set; } = [];
